backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pgvector
python-dotenv
openai
pydantic
pydantic-settings
requests
bcrypt==4.0.1
passlib[bcrypt]
python-jose[cryptography]
python-multipart
email-validator
asyncpg
pypdf
tiktoken